[UIMA-6372] Upgrade from JUnit 3 to JUnit 4 to JUnit 5
- Upgrade two additional test classes which had slipped through in the last refactoring iteration
